requirements.txt
jupyter
ipywidgets
numpy
tensorflow
matplotlib
seaborn
h5py
sklearn
keras
flask
pillow